backend/requirements.txt
Django==5.0.3
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9
Pillow==10.3.0
django-cors-headers==4.4.0
drf-yasg==1.21.7
python-dotenv==1.0.1



